--- Client/.dockerignore
@@ -1,4 +1,0 @@
-.git
-.gitignore
-node_modules
-dist
