--- Client/Dockerfile
@@ -1,18 +1,0 @@
-FROM node:12-alpine AS base
-WORKDIR /usr/src/app
-COPY package.json package-lock.json ./
-
-FROM base as build
-RUN apk add --no-cache --virtual /tmp/.gyp \
-    python \
-    make \
-    g++ \
-    && npm ci \
-    && apk del /tmp/.gyp
-COPY . .
-RUN  npm run build
-
-FROM base as release
-RUN npm install --production
-COPY --from=build /usr/src/app/dist /usr/src/app/dist
-CMD npm run start
