--- Client/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "client1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/index.ts",
-  "scripts": {
-    "pretest": "eslint src/**/*.ts",
-    "build": "tsc",
-    "prestart": "npm run build",
-    "start": "node ./dist"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@types/express": "^4.17.2",
-    "@typescript-eslint/eslint-plugin": "^2.19.0",
-    "typescript": "^3.7.5"
-  },
-  "dependencies": {
-    "@types/ws": "^7.2.1",
-    "@typescript-eslint/parser": "^2.19.0",
-    "eslint": "^6.8.0",
-    "express": "^4.17.1",
-    "ws": "^7.2.1"
-  }
-}
