+++ Client/requirements.txt
@@ -1,0 +1,5 @@
+adafruit-circuitpython-lis3dh
+adafruit-circuitpython-neopixel
+adafruit-blinka
+RPi.GPIO
+websockets
