--- Client/src/Client.ts
@@ -1,7 +1,0 @@
-export class Client {
-    constructor(username: string){
-        this.username = username;
-    }
-    
-    username: string;
-}
