--- PyClient/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "python.pythonPath": "/usr/bin/python",
-    "python.linting.pylintEnabled": true,
-    "python.linting.enabled": true
-}
