--- PyClient/active_data.txt
@@ -1,72 +1,0 @@
-Current colors
-(230.69331822809153, 44.34056851723457, 24.482329067636197)
-Current colors
-(230.69331822809153, 44.63795645300124, 24.89441226562813)
-Current colors
-(247.87711931999453, 44.63795645300124, 25.200761420290757)
-Current colors
-(232.70986174129308, 44.48876551360501, 24.964445284276113)
-Current colors
-(228.71142311958243, 44.63795645300124, 24.964445284276113)
-Current colors
-(234.76197028068245, 44.71292778140124, 24.82477107649156)
-Current colors
-(230.69331822809153, 44.41454339437742, 24.459835204341783)
-Current colors
-(38.504598611891204, 20.95567216849932, 34.69127036775986)
-Current colors
-(40.373185917897715, 73.86810903865951, 101.45600750534958)
-Current colors
-(27.952628469260556, 52.67525579323142, 28.18941472595751)
-Current colors
-(45.01534847822279, 31.64022106052622, 50.097996281648975)
-Current colors
-(112.61443431851887, 26.47909677251831, 62.143022413800495)
-Current colors
-(104.23670849078603, 22.804595891794282, 215.73755603009425)
-Current colors
-(69.43679740314458, 29.24218242607635, 32.931817820575155)
-Current colors
-(75.11869649665492, 33.724450513250474, 77.07604319579947)
-Current colors
-(39.012425694353745, 30.307470958115747, 125.93210682916286)
-Current colors
-(12.420495034204668, 168.06838329876118, 128.36087990370584)
-Current colors
-(26.03881460947501, 53.95637172187495, 215.73755603009425)
-Current colors
-(44.939360428625676, 23.08139859758994, 16.37879143788355)
-Current colors
-(33.68178295903862, 12.490423679573018, 28.07052225942886)
-Current colors
-(15.693284517095446, 13.098843144694772, 35.14930117779991)
-Current colors
-(71.29637772274361, 20.503763161404304, 24.193097722946963)
-Current colors
-(71.48782860200484, 29.371229957329767, 19.90585188296234)
-Current colors
-(24.70956427021025, 35.81124625914842, 207.3365848776978)
-Current colors
-(66.99061153098816, 33.26097127100628, 51.453565377674515)
-Current colors
-(36.69986234190829, 46.02710359367777, 23.324062601518225)
-Current colors
-(26.244167962674965, 30.941518197159542, 130.2448424676329)
-Current colors
-(30.727241846146892, 35.956348481605474, 130.88517977041369)
-Current colors
-(30.10185726491883, 37.740402269371984, 132.18493092420397)
-Current colors
-(29.83200657122634, 38.504598611891204, 135.5501214495857)
-Current colors
-(29.50142152651539, 39.889240293570445, 132.18493092420397)
-Current colors
-(29.08245896803357, 40.190336445334914, 149.22667506576562)
-Current colors
-(28.830499276411, 40.55770677410124, 155.32110031700097)
-Current colors
-(28.830499276411, 41.50620264781133, 144.37115300597824)
-Current colors
-(28.491099268449467, 41.70125048041949, 143.5924536147799)
-Current colors
-(28.309318612478624, 42.23045385853032, 140.5598839150146)
