--- PyClient/active_stdev.txt
@@ -1,12 +1,0 @@
-x std: 2.8803030731489336
-y std: 3.277102549502623
-z std: 3.4667277892715913
-x std: 2.6981988022134495
-y std: 5.601705120544591
-z std: 5.600460138873959
-x std: 3.98670822114201
-y std: 4.295675702674347
-z std: 2.7933109994415743
-x std: 1.8411271266806688
-y std: 2.5070113834601284
-z std: 4.219176665519729
