--- PyClient/idle_data.txt
@@ -1,54 +1,0 @@
-Current colors
-(243.34556013348495, 45.09159393975197, 24.941057204176815)
-Current colors
-(236.85059308763036, 44.71292778140124, 24.55005957514038)
-Current colors
-(224.8480664648925, 44.86362848958666, 24.437382636908787)
-Current colors
-(241.14135043312953, 44.266839648844595, 25.36886249540835)
-Current colors
-(228.71142311958243, 44.34056851723457, 24.847941472767577)
-Current colors
-(236.85059308763036, 44.12011504587503, 25.08204673824943)
-Current colors
-(232.70986174129308, 44.41454339437742, 24.941057204176815)
-Current colors
-(236.85059308763036, 44.71292778140124, 24.917712905567708)
-Current colors
-(245.59043777176112, 44.63795645300124, 24.89441226562813)
-Current colors
-(243.34556013348495, 44.71292778140124, 24.801643852170734)
-Current colors
-(250.20678345579, 44.34056851723457, 24.64095202694705)
-Current colors
-(230.69331822809153, 44.71292778140124, 24.755518437654576)
-Current colors
-(250.20678345579, 44.71292778140124, 24.917712905567708)
-Current colors
-(230.69331822809153, 44.34056851723457, 24.847941472767577)
-Current colors
-(234.76197028068245, 44.63795645300124, 24.847941472767577)
-Current colors
-(241.14135043312953, 44.71292778140124, 24.801643852170734)
-Current colors
-(243.34556013348495, 44.56323611650725, 25.034873448651787)
-Current colors
-(247.87711931999453, 44.56323611650725, 25.153140950375743)
-Current colors
-(243.34556013348495, 44.56323611650725, 24.847941472767577)
-Current colors
-(226.76329102488822, 44.41454339437742, 25.01135328253578)
-Current colors
-(241.14135043312953, 44.56323611650725, 24.917712905567708)
-Current colors
-(232.70986174129308, 44.788151368489316, 25.058437892173643)
-Current colors
-(238.97671348746604, 44.86362848958666, 24.89441226562813)
-Current colors
-(219.29170600746608, 44.71292778140124, 25.034873448651787)
-Current colors
-(250.20678345579, 44.41454339437742, 25.36886249540835)
-Current colors
-(232.70986174129308, 44.48876551360501, 24.847941472767577)
-Current colors
-(236.85059308763036, 44.19335556406504, 24.755518437654576)
