--- PyClient/idle_stdev.txt
@@ -1,15 +1,0 @@
-x std: 0.05840710024272803
-y std: 0.030127930101452902
-z std: 0.06108256448090045
-x std: 0.05332706559465588
-y std: 0.02860040476084097
-z std: 0.06358155898477279
-x std: 0.05054775948314551
-y std: 0.028982799021852076
-z std: 0.08805748945699618
-x std: 0.047037757607762964
-y std: 0.025812658823005187
-z std: 0.08233805487596124
-x std: 0.0547233970561494
-y std: 0.03284744809262495
-z std: 0.09190490723760741
