--- PyClient/requirements.txt
@@ -1,5 +1,0 @@
-adafruit-circuitpython-lis3dh
-adafruit-circuitpython-neopixel
-adafruit-blinka
-RPi.GPIO
-websockets
